lint

+++ pkg/ux/spinner.go
@@ -67,9 +67,9 @@ func (us *UserSpinner) SpinFailWithError(s *ysmrr.Spinner, txt string, err error
 		elapsed = time.Since(startTime)
 	}
 	if txt == "" {
-		s.UpdateMessage(fmt.Sprintf("%s err:%v [%.1fs]", s.GetMessage(), err, float64(elapsed.Seconds())))
+		s.UpdateMessage(fmt.Sprintf("%s err:%v [%.1fs]", s.GetMessage(), err, elapsed.Seconds()))
 	} else {
-		s.UpdateMessage(fmt.Sprintf("%s txt:%s err:%v [%.1fs]", s.GetMessage(), txt, err, float64(elapsed.Seconds())))
+		s.UpdateMessage(fmt.Sprintf("%s txt:%s err:%v [%.1fs]", s.GetMessage(), txt, err, elapsed.Seconds()))
 	}
 	s.Error()
 	Logger.log.Info(s.GetMessage() + " [Spinner Err]")
@@ -80,7 +80,7 @@ func (us *UserSpinner) SpinComplete(s *ysmrr.Spinner) {
 	if startTime, ok := us.startTimes[s]; ok {
 		elapsed = time.Since(startTime)
 	}
-	s.UpdateMessage(fmt.Sprintf("%s [%.1fs]", s.GetMessage(), float64(elapsed.Seconds())))
+	s.UpdateMessage(fmt.Sprintf("%s [%.1fs]", s.GetMessage(), elapsed.Seconds()))
 	s.Complete()
 	Logger.log.Info(s.GetMessage() + " [Spinner Complete]")
 }
